Getting results out of solr collection
